pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "rift-github-runner"
version = "0.1.0"
description = "Docker-compose deployable controller for ephemeral GitHub Actions runners on CloudRift VMs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3,<4",
    "requests>=2.28,<3",
    "sqlalchemy>=2,<3",
    "apscheduler>=3.10,<4",
    "gunicorn>=21,<23",
    "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7,<9",
    "responses>=0.23,<1",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["rift_github_runner*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
